# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Dokandar Mama shop assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: lowStockOnly
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Product" }
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
  /products/barcode/{code}:
    get:
      operationId: getProductByBarcode
      tags: [products]
      summary: Look up a product by barcode
      parameters:
        - name: code
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Found product
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
...
  /purchases:
    get:
      operationId: listPurchaseInvoices
      tags: [purchases]
      summary: Purchase invoice history
      responses:
        "200":
          description: Invoices
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/PurchaseInvoiceSummary" }
    post:
      operationId: createPurchaseInvoice
      tags: [purchases]
      summary: Submit invoice OCR text for parsing and matching
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PurchaseInvoiceInput" }
      responses:
        "201":
          description: Parsed invoice
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PurchaseInvoice" }
  /purchases/{id}:
    get:
      operationId: getPurchaseInvoice
      tags: [purchases]
      summary: Get a parsed invoice
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Invoice
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PurchaseInvoice" }
  /purchases/{id}/confirm:
    post:
      operationId: confirmPurchaseInvoice
      tags: [purchases]
      summary: Confirm invoice lines and update inventory
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PurchaseConfirmInput" }
      responses:
        "200":
          description: Applied
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PurchaseConfirmResult" }
...
components:
  schemas:
...
    PurchaseInvoiceInput:
      type: object
      required: [rawText]
      properties:
        rawText: { type: string, minLength: 1 }
        imageUrl: { type: string }
        supplierName: { type: string }
    PurchaseInvoiceSummary:
      type: object
      required: [id, status, createdAt]
      properties:
        id: { type: integer }
        supplierName: { type: ["string", "null"] }
        status: { type: string, enum: [uploaded, parsed, confirmed, failed] }
        invoiceTotal: { type: ["number", "null"] }
        createdAt: { type: string }
        confirmedAt: { type: ["string", "null"] }
    PurchaseInvoiceItem:
      type: object
      required: [id, name, quantity, status]
      properties:
        id: { type: integer }
        rawText: { type: ["string", "null"] }
        name: { type: string }
        quantity: { type: number }
        unitCost: { type: ["number", "null"] }
        lineTotal: { type: ["number", "null"] }
        matchedProductId: { type: ["integer", "null"] }
        status: { type: string, enum: [pending, matched, new, skipped] }
        suggestions:
          type: array
          items:
            type: object
            required: [productId, name, score]
            properties:
              productId: { type: integer }
              name: { type: string }
              score: { type: number }
    PurchaseInvoice:
      allOf:
        - $ref: "#/components/schemas/PurchaseInvoiceSummary"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/PurchaseInvoiceItem" }
    PurchaseConfirmInput:
      type: object
      required: [items]
      properties:
        supplierName: { type: string }
        items:
          type: array
          minItems: 1
          items:
            type: object
            required: [name, quantity]
            properties:
              name: { type: string, minLength: 1 }
              quantity: { type: number }
              unitCost: { type: ["number", "null"] }
              productId: { type: ["integer", "null"] }
              createNew: { type: boolean }
              skip: { type: boolean }
              sellPrice: { type: number }
              unit: { type: string }
              category: { type: string }
    PurchaseConfirmResult:
      type: object
      required: [id, status, updatedProducts, createdProducts]
      properties:
        id: { type: integer }
        status: { type: string }
        updatedProducts: { type: integer }
        createdProducts: { type: integer }
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
    Product:
      type: object
      required:
        [
          id,
          name,
          category,
          unit,
          price,
          stock,
          lowStockThreshold,
          isPriceVariable,
          updatedAt,
        ]
      properties:
        id: { type: integer }
        name: { type: string }
        barcode: { type: ["string", "null"] }
        category: { type: string }
        unit: { type: string }
        price: { type: number }
        costPrice: { type: ["number", "null"] }
        stock: { type: number }
        lowStockThreshold: { type: number }
        isPriceVariable: { type: boolean }
        updatedAt: { type: string }
    ProductInput:
      type: object
      required: [name, category, unit, price, stock]
      properties:
        name: { type: string, minLength: 1 }
        barcode: { type: string }
        category: { type: string }
        unit: { type: string }
        price: { type: number, minimum: 0 }
        costPrice: { type: number, minimum: 0 }
        stock: { type: number, minimum: 0 }
        lowStockThreshold: { type: number, minimum: 0 }
        isPriceVariable: { type: boolean }
    ProductUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        barcode: { type: string }
        category: { type: string }
        unit: { type: string }
        price: { type: number, minimum: 0 }
        costPrice: { type: number, minimum: 0 }
        stock: { type: number, minimum: 0 }
        lowStockThreshold: { type: number, minimum: 0 }
        isPriceVariable: { type: boolean }
...